feat: 일관성 있도록 타입 변경 및 예상 질문 response 일부 수정(#12)
회의 내용에 따라 예상 질문 목록 응답 시 labelId 대신 labenContent를 응답하도록 함

+++ src/main/java/reviewme/be/question/response/CommentOfQuestionResponse.java
@@ -38,5 +38,5 @@ public class CommentOfQuestionResponse {
     private List<Emoji> emojis;
 
     @Schema(description = "내가 선택한 이모지", example = "1")
-    private Long myEmojiId;
+    private Integer myEmojiId;
 }
